% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{build_model}
\alias{build_model}
\title{Machine Learning Modeling}
\usage{
build_model(
  x = NULL,
  y = NULL,
  test_x,
  daily_price_df,
  other_granularity_df,
  daily_price_df_date_col,
  other_granularity_df_date_col,
  independent_variables,
  dependent_variable,
  lag = 1,
  model_type = "rpart"
)
}
\arguments{
\item{x}{Data frame containing date column and independent variables with or without missing values}

\item{y}{Series of outcome variable values}

\item{daily_price_df}{Alternate data frame for modeling: daily price data frame containing date column and price}

\item{other_granularity_df}{Alternate data frame for modeling: independent variable data frame containing date column and independent variables}

\item{daily_price_df_date_col}{Name of date column in `daily_price_df`}

\item{other_granularity_df_date_col}{Name of date column in `other_granularity_df`}

\item{independent_variables}{Names of independent variables in `other_granularity_df`. All columns in `other_granularity_df` are considered if this is left as NULL}

\item{dependent_variable}{Name of dependent variable in `daily_price_df`}

\item{lag}{Number of days to lag the series for modeling. For example, if lag = 1 yesterday's independent variable values are used for predicting today's price}

\item{model_type}{Type of machine learning model to be built. Currently `rpart` and `lm` are supported}

\item{n}{Number of days to lead and lag around week ending. Prices are compared between `week_end + n` and `week_end - n`}
}
\value{
List containing the model, predictions for training set and predictions for test set along with residues
}
\description{
Builds machine learning model for a given granularity of data by merging with the daily price data. The inputs can be provided either as x, y, model_type, test_x where the data is considered as merged (independent variables may have missing values) or as
}
\examples{

}
