% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wasde_preprocessing.R
\name{clean_wasde}
\alias{clean_wasde}
\title{Clean the read-in WASDE soybean data}
\usage{
clean_wasde(com_data)
}
\arguments{
\item{com_data}{Data frame of the combined soybean data}
}
\value{
Data frame of the clean combined soybean data
}
\description{
Remove hifens and asterisk as well as transpose the data into time-series data
}
\examples{
data=read_wasde(path="C:/Users/40463/OneDrive/Documents/2019 Fall/Minne/WASDE")
}
