% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weekly_exploration.R
\name{create_corr_plot}
\alias{create_corr_plot}
\title{Correlation Plot}
\usage{
create_corr_plot(independent_var_names, dependent_var_names, df, ...)
}
\arguments{
\item{independent_var_names}{Character vector containing names of independent variables}

\item{dependent_var_names}{Character vector containing names of dependent variables}

\item{df}{Data frame containing price merged with exports}
}
\value{
Correlation matrix between set of independent variables and set of dependent variables
}
\description{
Creates correlation plot between dependent variables and independent variables for selected countries
}
\examples{
### Example 1 ###
data("soybeanExports")
competitors <- c("ARGENTINA", "BRAZIL")
df_total_export <- soybeanExports \%>\% group_by(Country) \%>\%
  summarize(Total_Export = sum(Weekly_Exports, na.rm = T))
top_countries <- head(x = df_total_export$Country[
  order(df_total_export$Total_Export, decreasing = TRUE)], n = 10)
selected_countries <- c(competitors, top_countries)
df_top_export <- soybeanExports[sapply(
  soybeanExports$Country, function(country) country \%in\% selected_countries), ]
df_top_export = inner_join(contractsForJuly2020, df_top_export, by = "Date")
ind_col <- c("Weekly_Exports", "CMY_Outstanding_Sales", "CMY_Gross_New_Sales",
             "CMY_Net_Sales" ,"CMY_Total_Commitment")
corr_mat_weekly = corr_mat_weekly <- create_corr_plot(
independent_var_names = ind_col,
dependent_var_names = colnames(df_top_export)[
  grep("Close",colnames(df_top_export))],
selected_countries = selected_countries,
country_var = "Country",
remove_countries = c("GRAND TOTAL", "KNOWN"),
df = df_top_export, top_corr = 4)

### Example 2 ###

}
