% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weekly_exploration.R
\name{create_export_corr_plot}
\alias{create_export_corr_plot}
\title{Correlation Plot}
\usage{
create_export_corr_plot(
  independent_var_names,
  dependent_var_names,
  selected_countries,
  df,
  country_var = "Country",
  top_corr = 4,
  remove_countries = c("GRAND TOTAL", "KNOWN")
)
}
\arguments{
\item{independent_var_names}{Character vector containing names of independent variables}

\item{dependent_var_names}{Character vector containing names of dependent variables}

\item{selected_countries}{Character vector containing names of selected countries}

\item{df}{Data frame containing price merged with exports}

\item{country_var}{Name of the column with Country}

\item{top_corr}{Specifies 'k' in the top-k correlations for each country}

\item{remove_countries}{Country names to be removed}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Creates correlation plot between dependent variables and independent variables for selected countries
}
\examples{
data("soybeanExports")
competitors <- c("ARGENTINA", "BRAZIL")
df_total_export <- soybeanExports \%>\% group_by(Country) \%>\%
  summarize(Total_Export = sum(Weekly_Exports, na.rm = T))
top_countries <- head(x = df_total_export$Country[
  order(df_total_export$Total_Export, decreasing = TRUE)], n = 10)
selected_countries <- c(competitors, top_countries)
df_top_export <- soybeanExports[sapply(
  soybeanExports$Country, function(country) country \%in\% selected_countries), ]
df_top_export = inner_join(contractsForJuly2020, df_top_export, by = "Date")
ind_col <- c("Weekly_Exports", "CMY_Outstanding_Sales", "CMY_Gross_New_Sales",
             "CMY_Net_Sales" ,"CMY_Total_Commitment")
corr_mat_weekly = create_export_corr_plot(independent_var_names = ind_col,
                                   dependent_var_names = colnames(df_top_export)[
                                   grep("Close",colnames(df_top_export))],
                                   selected_countries = selected_countries,
                                   df = df_top_export, top_corr = 4)
}
