% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_progress_util.R
\name{crop_progress_group_diff}
\alias{crop_progress_group_diff}
\title{Crop progress row-diff}
\usage{
crop_progress_group_diff(df)
}
\arguments{
\item{df}{Data frame containing raw crop progress data}
}
\value{
Data frame of diff
}
\description{
Computes row-wise diff for crop progress data
}
\examples{
diff_df_2017 <- crop_progress_group_diff(soybeanCropProgress2017)
}
