% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_preprocessing.R
\name{get_dtm}
\alias{get_dtm}
\title{Document Term Matrix}
\usage{
get_dtm(text, thr = 20, norm = F)
}
\arguments{
\item{full_df}{Path to a CSV file with export sales report}
}
\value{
Data frame: Document term matrix
}
\description{
Returns the document term matrix of given text input preprocessed using `FinancialModelingR`
}
\examples{
price_df <- read_exports("raw_data/ExportSalesDataByCommodity(Soybeans).csv", skip_lines = 4)
}
