% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{merge_price_other_df}
\alias{merge_price_other_df}
\title{Merge Different Granularities}
\usage{
merge_price_other_df(daily_price_df, other_granularity_df, lag = 1)
}
\arguments{
\item{daily_price_df}{Alternate data frame for modeling: daily price data frame containing date column and price}

\item{other_granularity_df}{Alternate data frame for modeling: independent variable data frame containing date column and independent variables}

\item{lag}{Number of days to lag the series for modeling. For example, if lag = 1 yesterday's independent variable values are used for predicting today's price}
}
\value{
Data frame with merged output of daily price data and other data frame of different granularity. The missing values in other data frame are filled with previous known value
}
\description{
Merges data frames of different time granularities and fills missing values with previous non-missing value
}
\examples{
merge_price_other_df(daily_price_df = contractsForJuly2020,
other_granularity_df = soybeanWASDE_clean,
lag = 1)
}
