% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly_trans.R
\name{monthly_trans}
\alias{monthly_trans}
\title{Combine monthly data and daily data and return monthly data}
\usage{
monthly_trans(
  daily_data,
  monthly_data,
  date_col1 = "Date",
  date_col2 = "Date"
)
}
\arguments{
\item{daily_data}{Data frame containing daily data}

\item{monthly_data}{Data frame containing monthly data}

\item{date_col1}{Name of the date column in daily data}

\item{date_col2}{Name of date column in monthly data}
}
\value{
Data frame of the combined final monthly data
}
\description{
Calculate the average price of from the day after the monthly observation day to day before next monthly observation day
}
\examples{
finaldata=monthly_trans(daily_data, monthly_data)
}
