% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress_exploration.R
\name{plot_price_vs_weekly_series}
\alias{plot_price_vs_weekly_series}
\title{Weekly Series vs Price Plot}
\usage{
plot_price_vs_weekly_series(
  df1_progress,
  df2_contracts,
  week_ending_col = "WEEK.ENDING",
  n = 2,
  line_columns = c("PROGRESS.in.PCT.PLANTED", "PROGRESS..5.YEAR.AVG.in.PCT.PLANTED",
    "PROGRESS..PREVIOUS.YEAR.in.PCT.PLANTED", "PROGRESS.in.PCT.EMERGED",
    "PROGRESS..5.YEAR.AVG.in.PCT.EMERGED", "PROGRESS..PREVIOUS.YEAR.in.PCT.EMERGED",
    "PROGRESS.in.PCT.BLOOMING", "PROGRESS..5.YEAR.AVG.in.PCT.BLOOMING",
    "PROGRESS..PREVIOUS.YEAR.in.PCT.BLOOMING", "PROGRESS.in.PCT.SETTING.PODS",
    "PROGRESS..5.YEAR.AVG.in.PCT.SETTING.PODS",
    "PROGRESS..PREVIOUS.YEAR.in.PCT.SETTING.PODS", "PROGRESS.in.PCT.DROPPING.LEAVES",
    "PROGRESS..5.YEAR.AVG.in.PCT.DROPPING.LEAVES",     
    "PROGRESS..PREVIOUS.YEAR.in.PCT.DROPPING.LEAVES", "PROGRESS.in.PCT.HARVESTED",
    "PROGRESS..5.YEAR.AVG.in.PCT.HARVESTED", "PROGRESS..PREVIOUS.YEAR.in.PCT.HARVESTED"),
  line_colors = c("Planted", "Planted", "Planted", "Emerged", "Emerged", "Emerged",
    "Blooming", "Blooming", "Blooming", "Setting Pods", "Setting Pods", "Setting Pods",
    "Dropping Leaves", "Dropping Leaves", "Dropping Leaves", "Harvested", "Harvested",
    "Harvested"),
  color_mapping = c(Planted = "brown", Emerged = "darkgreen", Blooming = "green",
    `Setting Pods` = "gold", `Dropping Leaves` = "red", Harvested = "purple"),
  line_type = c("Current Year", "5 Years Average", "Previous Year", "Current Year",
    "5 Years Average", "Previous Year", "Current Year", "5 Years Average",
    "Previous Year", "Current Year", "5 Years Average", "Previous Year", "Current Year",
    "5 Years Average", "Previous Year", "Current Year", "5 Years Average",
    "Previous Year"),
  color_variable_name = "Stage",
  variable_name = "Variable",
  variable_mapping = c(`Current Year` = 1, `5 Years Average` = 2, `Previous Year` = 3)
)
}
\arguments{
\item{df1_progress}{Data frame containing weekly series}

\item{df2_contracts}{Data frame containing close prices}

\item{week_ending_col}{Column name in `df1_progress` that contains the end of week for each data point}

\item{n}{Number of days to lead and lag around week ending. Prices are compared between `week_end + n` and `week_end - n`}

\item{line_columns}{Columns in `df1_progress` containing required weekly series to plot, corresponds to the exhaustive list of different types of series and variables among each series}

\item{line_colors}{Color sequence (strings) corresponding to the line_columns}

\item{color_mapping}{Mapping from string sequence of line colors (series type) to string sequence of predefined colors in R}

\item{line_type}{Line type sequence (strings) corresponding to each line added using line_columns}

\item{color_variable_name}{Proxy for each type of weekly series used in plotting - appears in legend}

\item{variable_name}{Proxy for each type of variable user in plotting - appears in legend}

\item{variable_mapping}{Line type mapping for different types of variables}
}
\value{
ggplot object with plots comparing price against progress over current year, previous year and 5 year average
}
\description{
Plots the multiple weekly data series vs daily price series. Prices are considered n days before and n days after the end date of each week. Line color corresponds to the type of weekly series and line type corresponds to the type of variable.
}
\examples{
contractsForJuly2020 <- read_price(
in_file = "private_data/ActiveSoybeanContractsforJuly2020.xlsx", delta_price = T,
add_delta = T, subset = T, subset_min_date = "2017-01-01",
rename_price_columns = T, rename_prefix = "july_2020_", skip_lines = 3)
plt_pretty(df1_progress = soybeanCropProgress2017, df2_contracts = contractsForJuly2020)
}
