% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wordcloud.R
\name{plot_wordcloud}
\alias{plot_wordcloud}
\title{Sentiment Wordcloud}
\usage{
plot_wordcloud(df1)
}
\arguments{
\item{df1}{Data frame with}
}
\value{
Wordcloud object based on word occurrence frequency colored by average sentiment of word
}
\description{
Plots wordcloud colored by sentiment of the preprocessed data obtained using `FinancialModelingR`
}
\examples{
df1 <- readRDS("raw_data/Tweets/unprocessed_China tweets @readDonaldTrump.csv.Rds")
plot_wordcloud(df1)
}
