% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress_exploration.R
\name{plt_pretty}
\alias{plt_pretty}
\title{Progress - Price Plot}
\usage{
plt_pretty(df1_progress, df2_contracts)
}
\arguments{
\item{df1_progress}{Data frame containing progress}

\item{df2_contracts}{Data frame containing close prices}
}
\value{
ggplot object with plots comparing price against progress over current year, previous year and 5 year average
}
\description{
Plots the weekly progress (in percentage) vs price series. Prices are considered 2 days before and 2 days after the end date of each week
}
\examples{
contractsForJuly2020 <- read_price(
in_file = "private_data/ActiveSoybeanContractsforJuly2020.xlsx", delta_price = T,
add_delta = T, subset = T, subset_min_date = "2017-01-01",
rename_price_columns = T, rename_prefix = "july_2020_", skip_lines = 3)
plt_pretty(df1_progress = soybeanCropProgress2017, df2_contracts = contractsForJuly2020)
}
