% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_preprocessing.R
\name{process_text}
\alias{process_text}
\title{Preprocess Tweets}
\usage{
process_text(file, processed = "processed", stopwords_file = NULL)
}
\arguments{
\item{file}{Location of Tweet CSV file}

\item{processed}{Parameter in c("processed", "unprocessed"). Performs cleaning based on part-of-speech and dependencies using spaCy and extracts sentiments if this is set to "processed". Otherwise it directly extracts sentiments on the whole Tweet}

\item{stopwords_file}{Text file containing stopwords - one in each line}
}
\value{
Data frame with sentiment summary of each Tweet in text file
}
\description{
Read and preprocess standard Tweet input CSV file
}
\examples{
df = process_text(path="raw_data/Tweets/China tweets @realDonaldTrump.csv")
}
