% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{read_exports}
\alias{read_exports}
\title{Read Export from CSV File}
\usage{
read_exports(file, skip_lines)
}
\arguments{
\item{file}{Path to a CSV file with export sales report}

\item{skip_lines}{Number of lines to skip on top of the file}
}
\value{
Data frame with Date column and export/sales columns
}
\description{
Reads an CSV file with Date and export/sales to different countries
}
\examples{
price_df <- read_exports("raw_data/ExportSalesDataByCommodity(Soybeans).csv", skip_lines = 4)
}
