% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wasde_util.R
\name{read_wasde}
\alias{read_wasde}
\title{Read in WASDE data}
\usage{
read_wasde(
  path,
  sheet_name = "Page 15",
  col_range = "A13:A58",
  remove_slice = c(3, 5, 17:21, 33:37)
)
}
\arguments{
\item{path}{Folder containing only all the Excel WASDE reports with proper file names}

\item{sheet_name}{Name of the sheet (string) with the required crop data}

\item{col_range}{Range of columns to read from the input file and sheet}

\item{remove_slice}{Indices of columns to remove from the column range. Use NULL to keep all columns}
}
\value{
Data frame of the combined data for required commodity (example: soybeans, corn)
}
\description{
Extract only crop specific WASDE reports across multiple WASDE files in a folder
}
\examples{
data=read_wasde(path="raw_data/WASDE")
}
