% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_plot.R
\name{time_plot}
\alias{time_plot}
\title{Visualize the time-series price plot}
\usage{
time_plot(date, price_data, main = NULL, xlab = "Date", ylab = "Price")
}
\arguments{
\item{date}{a data frame or a column of dataframe containing the date with date format}

\item{price_data}{the dataframe containing the price data}

\item{main}{title of the plot}

\item{xlab}{x label of the plot}

\item{ylab}{y lable of the plot}
}
\value{
a time-series plot
}
\description{
Plot the price-date plot
}
\examples{
plot=time_plot(wasde[,1],wasde[,2:7])
}
