% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wasde_util.R
\name{clean_wasde}
\alias{clean_wasde}
\title{Clean the read-in WASDE soybean data}
\usage{
clean_wasde(combined_data)
}
\arguments{
\item{combined_data}{Data frame of the combined soybean data}
}
\value{
Data frame of the clean combined soybean data
}
\description{
Remove hiphens and asterisk as well as transpose the data into time-series data
}
\examples{
data <- clean_wasde(combined_data = soybeanWASDE)
}
