% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_explore.R
\name{create_corr_plot}
\alias{create_corr_plot}
\title{Correlation plot}
\usage{
create_corr_plot(
  independent_var_names,
  dependent_var_names,
  df,
  remove_words = c("amp", "get", "want", "many", "will", "just", "want", "can",
    "realdonaldtrump", "made", "going", "must", "make"),
  text_correlates = c("china", "trade", "money", "deal", "tariffs", "economy",
    "currency", "china_jobs"),
  price_correlates = c("july_delta_Close", "may_delta_Close", "march_delta_Close",
    "july_delta_Close_sign", "may_delta_Close_sign", "march_delta_Close_sign")
)
}
\arguments{
\item{independent_var_names}{Character vector specifying the independent variable names}

\item{dependent_var_names}{Character vector specifying the dependent variable names}

\item{remove_words}{Words to be removed that were not removed as stopwords}

\item{text_correlates}{Text (independent) variables to consider for computing and plotting correlation}

\item{price_correlates}{Price (dependent) variables to consider for computing and plotting correlation}
}
\value{
Plot object containing the heatmap plot of correlation between independent variable(s) and dependent variable(s)
}
\description{
Generate correlation heatmap of given independent variable(s) and dependent variable(s) in a given data frame
}
\examples{
independent_var_names <- setdiff(
colnames(df), c(colnames(df)[grep(colnames(df), pattern = "_Close$")],
                 colnames(df)[grep(colnames(df), pattern = "_Open$")],
                 colnames(df)[grep(colnames(df), pattern = "_High$")],
                 colnames(df)[grep(colnames(df), pattern = "_Low$")],
                 colnames(df)[grep(colnames(df), pattern = "_sign$")],
                 "Date"))
dependent_var_names <- c(colnames(df)[grep(colnames(df), pattern = "_Close$")],
                         colnames(df)[grep(colnames(df), pattern = "_Open$")],
                         colnames(df)[grep(colnames(df), pattern = "_High$")],
                         colnames(df)[grep(colnames(df), pattern = "_Low$")],
                         colnames(df)[grep(colnames(df), pattern = "_sign$")])
create_corr_plot(independent_var_names, dependent_var_names, df)
}
