% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly_trans.R
\name{monthly_trans}
\alias{monthly_trans}
\title{Combine monthly data and daily data and return monthly data}
\usage{
monthly_trans(daily_data, monthly_data)
}
\arguments{
\item{daily_data}{the dataframe containing daily data with Date as the first column}

\item{monthly_data}{the dataframe containing monthly data with Date as the first column}
}
\value{
Data frame of the combined final monthly data
}
\description{
Calculate the average price of from the day after the monthly observation day to day before next monthly observation day
}
\examples{
finaldata=monthly_trans(daily_data,monthly_data)
}
