% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress_exploration.R
\name{plot_price_vs_weekly_series}
\alias{plot_price_vs_weekly_series}
\title{Progress - Price Plot}
\usage{
plot_price_vs_weekly_series(
  df1_progress,
  df2_contracts,
  week_ending_col = "WEEK.ENDING",
  n = 2,
  line_columns = c("PROGRESS.in.PCT.PLANTED", "PROGRESS..5.YEAR.AVG.in.PCT.PLANTED",
    "PROGRESS..PREVIOUS.YEAR.in.PCT.PLANTED", "PROGRESS.in.PCT.EMERGED",
    "PROGRESS..5.YEAR.AVG.in.PCT.EMERGED", "PROGRESS..PREVIOUS.YEAR.in.PCT.EMERGED",
    "PROGRESS.in.PCT.BLOOMING", "PROGRESS..5.YEAR.AVG.in.PCT.BLOOMING",
    "PROGRESS..PREVIOUS.YEAR.in.PCT.BLOOMING", "PROGRESS.in.PCT.SETTING.PODS",
    "PROGRESS..5.YEAR.AVG.in.PCT.SETTING.PODS",
    "PROGRESS..PREVIOUS.YEAR.in.PCT.SETTING.PODS", "PROGRESS.in.PCT.DROPPING.LEAVES",
    "PROGRESS..5.YEAR.AVG.in.PCT.DROPPING.LEAVES",     
    "PROGRESS..PREVIOUS.YEAR.in.PCT.DROPPING.LEAVES", "PROGRESS.in.PCT.HARVESTED",
    "PROGRESS..5.YEAR.AVG.in.PCT.HARVESTED", "PROGRESS..PREVIOUS.YEAR.in.PCT.HARVESTED"),
  line_colors = c("Planted", "Planted", "Planted", "Emerged", "Emerged", "Emerged",
    "Blooming", "Blooming", "Blooming", "Setting Pods", "Setting Pods", "Setting Pods",
    "Dropping Leaves", "Dropping Leaves", "Dropping Leaves", "Harvested", "Harvested",
    "Harvested"),
  line_type = c("Current Year", "5 Years Average", "Previous Year", "Current Year",
    "5 Years Average", "Previous Year", "Current Year", "5 Years Average",
    "Previous Year", "Current Year", "5 Years Average", "Previous Year", "Current Year",
    "5 Years Average", "Previous Year", "Current Year", "5 Years Average",
    "Previous Year"),
  color_variable_name = "Stage",
  color_mapping = c(Planted = "brown", Emerged = "darkgreen", Blooming = "green",
    `Setting Pods` = "gold", `Dropping Leaves` = "red", Harvested = "purple"),
  variable_name = "Variable",
  variable_mapping = c(`Current Year` = 1, `5 Years Average` = 2, `Previous Year` = 3)
)
}
\arguments{
\item{df1_progress}{Data frame containing progress}

\item{df2_contracts}{Data frame containing close prices}
}
\value{
ggplot object with plots comparing price against progress over current year, previous year and 5 year average
}
\description{
Plots the multiple weekly data series vs daily price series. Prices are considered n days before and n days after the end date of each week
}
\examples{
contractsForJuly2020 <- read_price(
in_file = "private_data/ActiveSoybeanContractsforJuly2020.xlsx", delta_price = T,
add_delta = T, subset = T, subset_min_date = "2017-01-01",
rename_price_columns = T, rename_prefix = "july_2020_", skip_lines = 3)
plt_pretty(df1_progress = soybeanCropProgress2017, df2_contracts = contractsForJuly2020)
}
