% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_preprocessing.R
\name{process_text}
\alias{process_text}
\title{Process Tweet Text}
\usage{
process_text(
  file,
  processed = "processed",
  csv_dir = "../Shane_s_lab/Data_Graduate/",
  stopwords_file = "../../Naveen_NLP/stopwords.txt"
)
}
\arguments{
\item{file}{Path to a csv file with Tweets}

\item{processed}{Prefix of the processed Rds file}

\item{csv_dir}{Base directory of CSV file}

\item{stopwords_file}{Location of text file containing stopwords}
}
\value{
A data frame with summary of polarity of text Tweets such as average polarity of sentences, minimum (most negative) polarity of sentences, maximum (most positive) polarity of sentences, standard deviation of polarity of sentences and useful text in the Tweet obtained by retaining nouns, adjectives and adverbs.
}
\description{
Creates a data frame with Tweet summaries such as average polarity of sentences, minimum (most negative) polarity of sentences, maximum (most positive) polarity of sentences, standard deviation of polarity of sentences (returns 0 if ther number of sentences is 1) and useful text in the Tweet obtained by retaining nouns, adjectives and adverbs.
}
\examples{
process_text(file = "", processed = "processed", csv_dir = "../Shane_s_lab/Data_Graduate/", stopwords_file = "../../Naveen_NLP/stopwords.txt")
}
