% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{read_price}
\alias{read_price}
\title{Read Price from Excel File}
\usage{
read_price(
  in_file,
  delta_price = F,
  add_delta = F,
  subset = F,
  start_date = "2017-01-01",
  skip_lines = 3,
  rename_price_columns = F,
  rename_prefix = NULL
)
}
\arguments{
\item{delta_price}{Flag indicating whether change in price should be calculated}

\item{subset}{Flag indicating whether only a subset of the data should be used}

\item{start_date}{Character string indicating the minimum date to be included in subset after reading the price file}

\item{skip_lines}{Number of lines above the header line in the Excel file}

\item{file}{Path to a Excel file with Date, Open, High, Low and Close prices}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Reads an Excel file with Date, Open, High, Low and Close prices of a stock or future contract
}
\examples{
price_df <- read_price(in_file, delta_price = F, subset = F, skip_lines = 3)
delta_price_df <- read_price(in_file, delta_price = F, subset = F, skip_lines = 3)
}
